nexus-microservices/api-gateway/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pyjwt>=2.8.0
python-multipart>=0.0.6
